pyproject.toml
[build-system]
requires = ["hatchling>=1.27"]
build-backend = "hatchling.build"

[project]
name = "openflux"
version = "0.1.0"
description = "Open standard for AI agent telemetry. One schema across every framework."
readme = "README.md"
license = "MIT"
license-files = ["LICENSE"]
requires-python = ">=3.12"
authors = [{ name = "Advit Arora" }]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries",
    "Topic :: System :: Monitoring",
]

[project.optional-dependencies]
openai = ["openai-agents>=0.1"]
langchain = ["langchain-core>=0.1"]
claude-agent-sdk = ["claude-agent-sdk>=0.1"]
autogen = [
    "autogen-agentchat>=0.4",
    "autogen-ext[openai]>=0.4",
]
crewai = [
    "crewai[tools]>=0.80",
]
google-adk = ["google-adk>=0.1"]
mcp = ["mcp>=1.0"]
bedrock = ["boto3>=1.28"]
all = [
    "openflux[openai]",
    "openflux[langchain]",
    "openflux[claude-agent-sdk]",
    "openflux[autogen]",
    "openflux[crewai]",
    "openflux[google-adk]",
    "openflux[mcp]",
    "openflux[bedrock]",
]

[project.scripts]
openflux = "openflux.cli:main"

[project.urls]
Homepage = "https://github.com/advitrocks9/openflux"
Documentation = "https://github.com/advitrocks9/openflux/docs"
Repository = "https://github.com/advitrocks9/openflux"

[tool.hatch.build.targets.wheel]
packages = ["src/openflux"]

[tool.pytest.ini_options]
testpaths = ["tests/unit", "tests/acceptance"]
pythonpath = [".", "tests/unit", "tests/acceptance"]
markers = [
    "integration: tests that hit I/O (SQLite, network, filesystem)",
    "acceptance: end-to-end acceptance tests requiring real API keys",
]

[dependency-groups]
dev = [
    "pytest>=8",
    "ruff>=0.9",
    "pyright>=1.1",
]

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = ["E", "W", "F", "I", "UP", "B", "SIM", "N", "ANN", "C4", "PTH"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["ANN"]
"src/openflux/adapters/**" = ["ANN401"]

[tool.pyright]
pythonVersion = "3.12"
typeCheckingMode = "strict"
exclude = [
    "src/openflux/adapters/autogen.py",
    "src/openflux/adapters/bedrock.py",
    "src/openflux/adapters/claude_agent_sdk.py",
    "src/openflux/adapters/crewai.py",
    "src/openflux/adapters/google_adk.py",
    "src/openflux/adapters/langchain.py",
    "src/openflux/adapters/mcp.py",
    "src/openflux/adapters/openai_agents.py",
]
